sessition delete error fixed

+++ services/sesssionServices.js
@@ -48,7 +48,13 @@ export async function updateSession(sessionId, sessionData, updatedBy) {
   return await sessionCreationService.updateSession(sessionId, sessionData);
 }
 
-export async function deleteSession(sessionId) {
+export async function deleteSession(sessionId, instituteId, universityId) {
+  const isSessionAlreadyMapped = await sessionCreationService.isSessionMappedwithcourse(sessionId, instituteId, universityId);
+
+  if (isSessionAlreadyMapped.length > 0) {
+    throw new Error('Session already mapped with courses unable to delete')
+  } 
+
   return await sessionCreationService.deleteSession(sessionId);
 };
